Engine/requirements.txt
Flask>=2.2.3
flask-session>=0.5.0
flask-mail>=0.9.1
mysql-connector-python>=8.2.0
requests>=2.31.0
flask-cors>=4.0.0
flask-socketio>=5.1.0
APScheduler>=3.8.1